Expose Data.Kind and GHC.TypeNats
I've been using `rerebase` as my Prelude for `ghci` and personal projects for quite a while, but I just ran into the lack of `Data.Kind` when writing my own kinds. If this works for you, I'll open a PR for `rerebase`, too (and push a version bump for both).